Move the em_table_lib.lua helper file elsewhere because we're going to need it.
