+++ h/em_table_lib.lua
